[LOOSECOUPLING]: router folder restructured and removed inline source map for smaller bundle size
