+++ src/lib/decorators.ts
@@ -2,7 +2,7 @@
 //https://jsfiddle.net/r5umxasz/
 import { Reflection as Reflect } from '@abraham/reflection';
 import { registerElement } from "./registerElement";
-import { Injector } from "./service_resolver";
+import { Injector } from "./router/service_resolver";
 import { INPUT_METADATA_KEY } from "./utils";
 import { DecoratorOptions } from "./types";
