+++ src/lib/router/index.ts
@@ -1,0 +1,2 @@
+export { registerRouterComponent } from './router';
+export { Router } from './routerService';
