+++ src/lib/router/routerService.ts
@@ -1,18 +1,16 @@
 //https://krasimirtsonev.com/blog/article/A-modern-JavaScript-router-in-100-lines-history-api-pushState-hash-url
-import { isArray } from "./utils";
-import { Route, ICurrentRoute } from "./types";
-import { registerRouterComponent } from "./router";
+import { isArray } from "../utils";
+import { Route, ICurrentRoute } from "../types";
 import { isNode } from "browser-or-node";
 import { StaticRouter } from './staticRouter';
-import { Injectable } from './decorators';
-import { InternalRouter } from './internalRouterService';
+import { Injectable } from '../decorators';
+import { InternalRouter } from '../internalRouterService';
 
 @Injectable()
 export class Router {
 	getCurrentRoute: () => ICurrentRoute;
 	navigateTo: (path: string) => void;
 	constructor(internalRouter: InternalRouter) {
-		registerRouterComponent();
 		this.getCurrentRoute = internalRouter.getCurrentRoute.bind(internalRouter);
 		this.navigateTo = internalRouter.navigateTo.bind(internalRouter);
 	}
