+++ src/plume.ts
@@ -1,12 +1,12 @@
 //https://github.com/ibhi/webcomponent-with-di/blob/master/src/users.component.js
 //https://medium.com/@gilfink/creating-a-custom-element-decorator-using-typescript-302e7ed3a3d1
 "use strict";
-import { Injector } from "./lib/service_resolver";
+import { Injector } from "./lib/router/service_resolver";
 import { Component, Injectable, Input } from "./lib/decorators";
 import { DecoratorOptions, Route, Ref, IHooks } from "./lib/types";
 import { html } from "lighterhtml";
 import { useRef, useState, useContext } from "augmentor";
-import { Router } from "./lib/routerService";
+import { Router, registerRouterComponent } from "./lib/router";
 import { DomTransition } from "./lib/domTransition.service";
 import { TranslationService } from "./lib/translationService";
 import { useFormFields } from './lib/hooks';
@@ -17,6 +17,7 @@ export {
   Injector,
   Input,
   Router,
+  registerRouterComponent,
   TranslationService,
   DomTransition,
   DecoratorOptions,
